+++ src/java/co/edu/ufps/Sisvencat/models/ClasesDAO/PedidoDAO.java
@@ -62,7 +62,7 @@ public class PedidoDAO implements Serializable, IDAOPedido{
         }
         
         PreparedStatement state = con.getConexion().prepareStatement(consulta);
-        state.setInt(1, cam.getCodigo_cam());
+        state.setLong(1, cam.getCodigo_cam());
         ResultSet rs = state.executeQuery();
         
         Pedido p = null;
